+++ src/command/Command.java
@@ -2,7 +2,7 @@ package command;
 
 public abstract class Command {
 
-    public abstract void execute(int a, int b);
+    public abstract void execute(double a, double b);
 
 }
